+++ saga-sample/spring-seata-dubbo-saga/e2e-files/commit.yaml
@@ -1,0 +1,1 @@
+{"res": "commit"}
