+++ saga-sample/spring-seata-dubbo-saga/e2e-files/rollback.yaml
@@ -1,0 +1,1 @@
+{"res": "rollback"}
