+++ saga-sample/spring-seata-dubbo-saga/spring-seata-dubbo-saga-consumer/src/main/java/org/apache/seata/consumer/config/DubboConsumerConfiguration.java
@@ -1,0 +1,71 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.seata.consumer.config;
+
+import org.apache.dubbo.config.ApplicationConfig;
+import org.apache.dubbo.config.ProtocolConfig;
+import org.apache.dubbo.config.RegistryConfig;
+import org.apache.dubbo.config.annotation.DubboReference;
+import org.apache.seata.provider.action.BalanceAction;
+import org.apache.seata.provider.action.InventoryAction;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import static org.apache.seata.e2e.E2EUtil.isInE2ETest;
+
+/**
+ * @author wangte
+ * Create At 2024/1/20
+ */
+@Configuration
+public class DubboConsumerConfiguration {
+
+    public static final String CONSUMER_APP_NAME = "seata-saga-consumer";
+
+    @DubboReference(id = "balanceAction")
+    private BalanceAction balanceAction;
+
+    @DubboReference(id = "inventoryAction")
+    private InventoryAction inventoryAction;
+
+    @Bean
+    public ApplicationConfig applicationConfig() {
+        ApplicationConfig applicationConfig = new ApplicationConfig(CONSUMER_APP_NAME);
+        applicationConfig.setQosEnable(false);
+        return applicationConfig;
+    }
+
+    @Bean
+    public RegistryConfig registryConfig() {
+        RegistryConfig registryConfig = new RegistryConfig();
+        if (isInE2ETest()) {
+            registryConfig.setAddress("zookeeper://zookeeper:2181");
+        } else {
+            registryConfig.setAddress("zookeeper://localhost:2181");
+        }
+        return registryConfig;
+    }
+
+    @Bean
+    public ProtocolConfig protocolConfig() {
+        ProtocolConfig protocolConfig = new ProtocolConfig();
+        protocolConfig.setName("dubbo");
+        protocolConfig.setPort(20882);
+        return protocolConfig;
+    }
+
+}
