+++ saga-sample/spring-seata-dubbo-saga/spring-seata-dubbo-saga-consumer/src/main/resources/statelang/reduce_inventory_and_balance.json
