+++ saga-sample/spring-seata-dubbo-saga/spring-seata-dubbo-saga-provider/src/main/java/org/apache/seata/provider/action/BalanceAction.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.seata.provider.action;
+
+import java.math.BigDecimal;
+import java.util.Map;
+
+/**
+ * Balance Actions
+ */
+public interface BalanceAction {
+
+    /**
+     * reduce
+     *
+     * @param businessKey
+     * @param amount
+     * @param params
+     * @return
+     */
+    boolean reduce(String businessKey, BigDecimal amount, Map<String, Object> params);
+
+    /**
+     * compensateReduce
+     *
+     * @param businessKey
+     * @param params
+     * @return
+     */
+    boolean compensateReduce(String businessKey, Map<String, Object> params);
+
+}
