+++ saga-sample/spring-seata-dubbo-saga/spring-seata-dubbo-saga-provider/src/main/resources/file.conf
@@ -1,0 +1,126 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+transport {
+  # tcp, unix-domain-socket
+  type = "TCP"
+  #NIO, NATIVE
+  server = "NIO"
+  #enable heartbeat
+  heartbeat = true
+  # the tm client batch send request enable
+  enableTmClientBatchSendRequest = false
+  # the rm client batch send request enable
+  enableRmClientBatchSendRequest = true
+   # the rm client rpc request timeout
+  rpcRmRequestTimeout = 2000
+  # the tm client rpc request timeout
+  rpcTmRequestTimeout = 30000
+  # the rm client rpc request timeout
+  rpcRmRequestTimeout = 15000
+  #thread factory for netty
+  threadFactory {
+    bossThreadPrefix = "NettyBoss"
+    workerThreadPrefix = "NettyServerNIOWorker"
+    serverExecutorThread-prefix = "NettyServerBizHandler"
+    shareBossWorker = false
+    clientSelectorThreadPrefix = "NettyClientSelector"
+    clientSelectorThreadSize = 1
+    clientWorkerThreadPrefix = "NettyClientWorkerThread"
+    # netty boss thread size
+    bossThreadSize = 1
+    #auto default pin or 8
+    workerThreadSize = "default"
+  }
+  shutdown {
+    # when destroy server, wait seconds
+    wait = 3
+  }
+  serialization = "seata"
+  compressor = "none"
+}
+service {
+  #transaction service group mapping
+  vgroupMapping.my_test_tx_group = "default"
+  #only support when registry.type=file, please don't set multiple addresses
+  default.grouplist = "127.0.0.1:8091"
+  #degrade, current not support
+  enableDegrade = false
+  #disable seata
+  disableGlobalTransaction = false
+}
+
+client {
+  rm {
+    asyncCommitBufferLimit = 10000
+    lock {
+      retryInterval = 10
+      retryTimes = 30
+      retryPolicyBranchRollbackOnConflict = true
+    }
+    reportRetryCount = 5
+    tableMetaCheckEnable = false
+    tableMetaCheckerInterval = 60000
+    reportSuccessEnable = false
+    sagaBranchRegisterEnable = false
+    sagaJsonParser = "fastjson"
+    sagaRetryPersistModeUpdate = false
+    sagaCompensatePersistModeUpdate = false
+    tccActionInterceptorOrder = -2147482648 #Ordered.HIGHEST_PRECEDENCE + 1000
+    sqlParserType = "druid"
+    branchExecutionTimeoutXA = 60000
+    connectionTwoPhaseHoldTimeoutXA = 10000
+  }
+  tm {
+    commitRetryCount = 5
+    rollbackRetryCount = 5
+    defaultGlobalTransactionTimeout = 60000
+    degradeCheck = false
+    degradeCheckPeriod = 2000
+    degradeCheckAllowTimes = 10
+    interceptorOrder = -2147482648 #Ordered.HIGHEST_PRECEDENCE + 1000
+  }
+  undo {
+    dataValidation = true
+    onlyCareUpdateColumns = true
+    logSerialization = "jackson"
+    logTable = "undo_log"
+    compress {
+      enable = true
+      # allow zip, gzip, deflater, lz4, bzip2, zstd default is zip
+      type = zip
+      # if rollback info size > threshold, then will be compress
+      # allow k m g t
+      threshold = 64k
+    }
+  }
+  loadBalance {
+      type = "XID"
+      virtualNodes = 10
+  }
+}
+log {
+  exceptionRate = 100
+}
+tcc {
+  fence {
+    # tcc fence log table name
+    logTableName = tcc_fence_log
+    # tcc fence log clean period
+    cleanPeriod = 1h
+  }
+}
